requirements.txt
torch>=2.0.0
transformers>=4.36.0
torchvision>=0.15.0
pillow>=9.3.0
tqdm>=4.65.0
numpy>=1.24.0
accelerate>=0.25.0 